Extracted URL extension into its own file
